requirements.txt
pillow~=10.1.0
transformers~=4.35.0
tqdm~=4.66.1
torchmetrics~=1.2.0
requests~=2.31.0
matplotlib~=3.8.1
torch~=2.1.0
torchvision~=0.16.0
numpy~=1.26.1